rename stackParser.js to extactPluginName.js move it to lib
