+++ lib/extractPluginName.js
